+++ src/contour/main.cpp
@@ -96,6 +96,10 @@ int main(int argc, char* argv[])
 
     try
     {
+        auto const HTS = "\033H";
+        auto const TBC = "\033[g";
+        printf("\r%s        %s                        %s\r", TBC, HTS, HTS);
+
         QCoreApplication::setApplicationName("contour");
         QCoreApplication::setOrganizationName("contour");
         QCoreApplication::setApplicationVersion(CONTOUR_VERSION_STRING);
@@ -204,6 +208,7 @@ int main(int argc, char* argv[])
         controller.exit();
         controller.wait();
 
+        printf("\r%s", TBC);
         return rv;
     }
     catch (exception const& e)